--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -108,7 +108,7 @@ services:
       # dockerfile to use for build
       dockerfile: Dockerfile
     # update version number to correspond to frontend/package.json
-    image: rsd/frontend:2.0.0
+    image: rsd/frontend:2.0.1
     environment:
       # it uses values from .env file
       - POSTGREST_URL