--- a/frontend/components/AppHeader/index.tsx
+++ b/frontend/components/AppHeader/index.tsx
@@ -64,8 +64,20 @@ export default function AppHeader() {
         className="flex-1 flex flex-col px-4 xl:flex-row items-start lg:container lg:mx-auto">
         <div className="w-full flex-1 flex items-center justify-between">
           <Link href="/" passHref className="hover:text-inherit" aria-label="Link to home page">
-            <LogoApp className="hidden 2xl:block"/>
-            <LogoAppSmall className="block 2xl:hidden"/>
+            <LogoApp
+              className="hidden 2xl:block"
+              loading='eager'
+              // lighthouse audit requires explicit width and height
+              width="100%"
+              height="1.5rem"
+            />
+            <LogoAppSmall
+              className="2xl:hidden"
+              loading='eager'
+              // lighthouse audit requires explicit width and height
+              width="100%"
+              height="1.5rem"
+            />
           </Link>
 
           <GlobalSearchAutocomplete className="hidden xl:block ml-12 mr-6"/>