--- a/frontend/components/projects/edit/impact/EditProjectImpactIndex.test.tsx
+++ b/frontend/components/projects/edit/impact/EditProjectImpactIndex.test.tsx
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2022 - 2023 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 - 2023 Dusan Mijatovic (dv4all) (dv4all)
 // SPDX-FileCopyrightText: 2022 - 2023 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 // SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
 // SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 //
@@ -21,7 +22,7 @@ import projectState from '../__mocks__/editProjectState'
 
 const mockGetImpactForProject = jest.fn((props) => Promise.resolve(mockImpactForProject))
 jest.mock('~/utils/getProjects', () => ({
-  getImpactForProject: jest.fn((props)=>mockGetImpactForProject(props))
+  getMentionsForProject: jest.fn((props)=>mockGetImpactForProject(props))
 }))
 
 const mockGetMentionByDoiFromRsd = jest.fn((props) => Promise.resolve([] as any))