--- a/frontend/components/projects/edit/impact/ImportProjectImpact.tsx
+++ b/frontend/components/projects/edit/impact/ImportProjectImpact.tsx
@@ -1,3 +1,4 @@
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 // SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
 // SPDX-FileCopyrightText: 2023 Netherlands eScience Center
@@ -10,7 +11,7 @@ import ImportMentions from '~/components/mention/ImportMentions/index'
 import ImportMentionsInfoPanel from '~/components/mention/ImportMentions/ImportMentionsInfoPanel'
 import useEditMentionReducer from '~/components/mention/useEditMentionReducer'
 import useProjectContext from '~/components/projects/edit/useProjectContext'
-import {getImpactForProject} from '~/utils/getProjects'
+import {getMentionsForProject} from '~/utils/getProjects'
 import {cfgImpact as config} from './config'
 
 export default function ImportProjectImpact() {
@@ -20,7 +21,7 @@ export default function ImportProjectImpact() {
 
   async function reloadImpact() {
     setLoading(true)
-    const data = await getImpactForProject({project: project.id, token: token, frontend: true})
+    const data = await getMentionsForProject({project: project.id,table:'impact_for_project',token: token})
     setMentions(data)
     setLoading(false)
   }