--- a/frontend/components/software/edit/mentions/useMentionForSoftware.tsx
+++ b/frontend/components/software/edit/mentions/useMentionForSoftware.tsx
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: 2022 - 2023 dv4all
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 // SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all) (dv4all)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -20,13 +22,12 @@ export default function useMentionForSoftware({software, token}: MentionForSoftw
 
   useEffect(() => {
     let abort = false
-    async function getImpactFromApi() {
+    async function getMentionsFromApi() {
       setLoading(true)
       // TODO! this request is made two times, investigate
       const mentionsForSoftware = await getMentionsForSoftware({
         software,
-        token,
-        frontend: true
+        token
       })
       if (mentionsForSoftware && abort === false) {
         // debugger
@@ -37,7 +38,7 @@ export default function useMentionForSoftware({software, token}: MentionForSoftw
     }
     if (software && token &&
       software !== loadedSoftware) {
-      getImpactFromApi()
+      getMentionsFromApi()
     }
 
     return () => { abort = true }