--- a/frontend/components/software/overview/list/SoftwareListItemContent.tsx
+++ b/frontend/components/software/overview/list/SoftwareListItemContent.tsx
@@ -5,7 +5,7 @@
 
 import {JSX} from 'react'
 import {getImageUrl} from '~/utils/editImage'
-import useValidateImageSrc from '~/utils/useValidateImageSrc'
+import ListImageWithGradientPlaceholder from '~/components/projects/overview/list/ListImageWithGradientPlaceholder'
 import SoftwareMetrics from '../cards/SoftwareMetrics'
 
 type SoftwareOverviewListItemProps = {
@@ -27,20 +27,13 @@ type SoftwareOverviewListItemProps = {
 
 export default function SoftwareListItemContent(item:SoftwareOverviewListItemProps) {
   const imgSrc = getImageUrl(item.image_id ?? null)
-  const validImg = useValidateImageSrc(imgSrc)
+
   return (
     <>
-      {validImg ?
-        <img
-          src={`${imgSrc ?? ''}`}
-          alt={`Cover image for ${item.brand_name}`}
-          className="w-12 max-h-[3.5rem] text-base-content-disabled p-2 object-contain object-center"
-        />
-        :
-        <div
-          className="w-12 self-stretch bg-gradient-to-br from-base-300 from-0% via-base-100 via-50% to-base-100"
-        />
-      }
+      <ListImageWithGradientPlaceholder
+        imgSrc={imgSrc}
+        alt = {`Cover image for ${item.brand_name}`}
+      />
       <div className="flex flex-col md:flex-row gap-3 flex-1 py-2">
         <div className="flex-1">
           <div className='line-clamp-2 md:line-clamp-1 break-words font-medium'>