--- a/frontend/pages/api/fe/auth/index.ts
+++ b/frontend/pages/api/fe/auth/index.ts
@@ -1,7 +1,8 @@
+// SPDX-FileCopyrightText: 2022 - 2023 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2022 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -35,6 +36,9 @@ export type Provider = {
 
 type Data = Provider[] | ApiError
 
+// cached list of providers
+let loginProviders:Provider[] = []
+
 async function getRedirectInfo(provider: string) {
   // select provider
   switch (provider.toLocaleLowerCase()) {
@@ -56,11 +60,9 @@ async function getRedirectInfo(provider: string) {
   }
 }
 
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<Data>
-) {
-  try {
+async function getProvidersInfo(){
+  // only if we did not loaded info previously
+  if (loginProviders.length === 0){
     // extract list of providers, default value surfconext
     const strProviders = process.env.RSD_AUTH_PROVIDERS || 'surfconext'
     // split providers to array on ;
@@ -82,8 +84,25 @@ export default async function handler(
         info.push(item.value as Provider)
       }
     })
+    // save response into cached variable
+    loginProviders = [
+      ...info
+    ]
+    return loginProviders
+  }
+  // console.log("getProvidersInfo...cached...loginProviders")
+  return loginProviders
+}
+
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse<Data>
+) {
+  try {
+    // extract list of providers from .env file
+    const providers = await getProvidersInfo()
     // return only 'valid' providers
-    res.status(200).json(info as Provider[])
+    res.status(200).json(providers)
   } catch (e: any) {
     logger(`api/fe/auth/index: ${e?.message}`, 'error')
     res.status(500).json({