… frontend/pages/projects/[slug]/index.tsx
… frontend/pages/software/[slug]/index.tsx
… frontend/utils/editMentions.ts
… frontend/utils/getProjects.ts